hw13/pyproject.toml
[tool.poetry]
name = "hw13"
version = "0.1.0"
description = ""
authors = ["Vaipik <n.khylko@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
django-autoslug = "^1.9.8"
psycopg2 = "^2.9.5"
pytils = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
